Added new command
Added loader command, this command will reload either all commands, or all events from their respective directories.
Updated update command, will now reboot if a restart flag is present

+++ commands/owner/update.js
@@ -23,7 +23,7 @@ class Update extends Owner {
     if (stdout) out.push(stdout);
     if (stderr) out.push(stderr);
     await message.channel.send(out.join("---\n"), { code: true });
-    if (!stdout.toString().includes("Already up-to-date.")) {
+    if (!stdout.toString().includes("Already up-to-date.") && message.flags[0] === "restart") {
       this.client.commands.get("reboot").run(message, args, level);
     }
   }
